check task jobform added, functioning
