+++ src/Components/Job/Functions/JobFormFunctions.js
@@ -17,8 +17,36 @@ function removeTask(stateVar, setFunction, index) {
     setFunction(remove)
 }
 
+function checkForm(obj, stateVar) {
+    const { jobDetails } = obj;
+    // key values
+    const originalForm = Object.values(stateVar);
+    const updatedForm = Object.values(jobDetails);
+
+    for (let i = 0; i < updatedForm.length; i++) {
+      if ((i === 8 || i === 6) && originalForm[i].length !== updatedForm[i].length) {
+        return true;
+      }
+      if(i === 6 || i === 8){
+        const originalArr = originalForm[i];
+        const updatedArr = updatedForm[i];
+        const changedArr = updatedArr.every((el) =>
+        originalArr.includes(el)
+        );
+        if (!changedArr) {
+        return true;
+        }
+      }
+      if (i !== 6 && i !== 8 && updatedForm[i] !== originalForm[i]) {
+        return true;
+      }
+    }
+    return false;
+  }
+
 export {
     handleTextChange,
     handleTasks,
     removeTask,
+    checkForm,
 }
